tidied up backend code and improved filters
